Split extra.calibration into its own package
